requirements.txt
fastapi==0.121.3
uvicorn==0.22.0
torch==2.2.0
torchvision==0.17.0
pytorch-lightning==2.1.2
pillow==10.0.0
openai==2.8.1
pydantic==2.12.4
requests==2.31.0
python-multipart==0.0.6
torchmetrics==1.2.1
numpy==1.26.4

langchain==0.2.11
langchain-core==0.2.23
langsmith==0.1.67
langchain-community==0.4.1
langchain-openai==1.1.0
lancedb==0.25.3
tiktoken==0.12.0
dotenv-python==0.0.1
langchain-text-splitters==1.0.0
BeautifulSoup==3.2.2